requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Cython==3.0.12
cython_bbox==0.1.5
debugpy==1.6.4
decorator==5.1.1
deep-sort-realtime==1.3.2
Django==5.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
entrypoints==0.4
executing==1.2.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.57.0
fsspec==2024.6.1
idna==3.10
ipykernel==6.19.4
ipython==8.7.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.8
lap==0.5.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
motmetrics==1.4.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.3
norfair==2.3.0
numpy==1.26.4
opencv-python==4.11.0.86
packaging==22.0
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==11.0.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.14.0
pyparsing==3.2.3
python-dateutil==2.8.2
pytz==2025.2
pywin32==305
PyYAML==6.0.2
pyzmq==24.0.1
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.2
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.2
tqdm==4.67.1
traitlets==5.8.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.120
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.5
xmltodict==0.14.2
